requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
# Database
psycopg2-binary==2.9.9
# WSGI Server
gunicorn==21.2.0
# Environment variables
python-dotenv==1.0.0
# Security
Flask-WTF==1.1.1
# Rate limiting
Flask-Limiter==3.5.0
# Caching
Flask-Caching==2.1.0
# Scheduler
schedule==1.2.0

# Excel Export
pandas==3.0.0
openpyxl==3.1.5
